% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhal.R
\name{cv_hal_usmooth}
\alias{cv_hal_usmooth}
\title{Undersmoothed Highly Adaptive Lasso for Cross-Validation}
\usage{
cv_hal_usmooth(
  fold,
  data_in,
  outcome_type,
  x_names,
  y_names,
  L1_seq = seq(0.2, 2000, length.out = 10000),
  lambda_seq = exp(seq(-0.5, -20, length = 10000)),
  basis_list = NULL
)
}
\arguments{
\item{fold}{Object specifying cross-validation folds as generated by a call
to \code{\link[origami]{make_folds}}.}

\item{data_in}{An input data set, usually a \code{data.table}.}

\item{outcome_type}{A \code{character} specifying the error family of the
regression model, passed directly to \code{\link[hal9001]{fit_hal}}.}

\item{x_names}{A \code{character} giving the name(s) of the column(s) in the
input dataset corresponding to the regressors/predictors.}

\item{y_names}{A \code{character} giving the name of the column in the input
dataset corresponding to the outcome.}

\item{L1_seq}{A \code{numeric} sequence of L1-norm values to be matched to
the corresponding values of the regularization parameter lambda.}

\item{lambda_seq}{A \code{numeric} sequence of values of the regularization
parameter, to be used in fitting HAL regressions and matched up to values
of the L1 norm.}

\item{basis_list}{A \code{list} of basis functions produced by a prior call
to \code{\link[hal9001]{fit_hal}}, used to ensure that basis functions are
preserved across a sequence of undersmoothed HAL regressions.}
}
\description{
Undersmoothed Highly Adaptive Lasso for Cross-Validation
}
