% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhal.R
\name{fit_gn_ipw}
\alias{fit_gn_ipw}
\title{Fit propensity score g_n(A|W) for IPW estimate of treatment-specific mean}
\usage{
fit_gn_ipw(data_in, folds, x_names, y_names, L1_seq, lambda_seq)
}
\arguments{
\item{data_in}{An input data set, usually a \code{data.table}.}

\item{folds}{A \code{list} of \code{Fold} objects specifying the full set of
cross-validation folds, generated by \code{\link[origami]{make_folds}}.}

\item{x_names}{A \code{character} giving the name(s) of the column(s) in the
input dataset corresponding to the regressors/predictors. For the treatment
mechanism, this should be only the baseline covariates W.}

\item{y_names}{A \code{character} giving the name of the column in the input
dataset corresponding to the outcome. For the treatment mechanism, this
should be only the treatment A.}

\item{L1_seq}{A \code{numeric} sequence of L1-norm values to be matched to
the corresponding values of the regularization parameter lambda.}

\item{lambda_seq}{A \code{numeric} sequence of values of the regularization
parameter, to be used in fitting HAL regressions and matched up to values
of the L1 norm.}
}
\description{
Fit propensity score g_n(A|W) for IPW estimate of treatment-specific mean
}
