% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npipw.R
\name{ipw_dcar}
\alias{ipw_dcar}
\title{Compute the Coarsened-at-Random Projection of IPW Estimators}
\usage{
ipw_dcar(a, y, gn, gn_dcar = NULL, Qn_tsm, est_type = c("ht", "hajek"))
}
\arguments{
\item{a}{Observed values of the treatment.}

\item{y}{Observed values of the outcome.}

\item{gn}{Estimates of the propensity score gn used in computing the score
(i.e., (a - gn)) for the treatment mechanism. These same estimates may be
reused in computing the EIF or be replaced instead by \code{gn_dcar}.}

\item{gn_dcar}{Estimates of the propensity score used in the re-weighting
term (i.e., Qn / gn) of the EIF. The default is \code{NULL}, in which case
\code{gn} is used in the re-weighting term as well as in the score.}

\item{Qn_tsm}{Counterfactual outcomes, predicted Qn or true Q0, used in the
score equation D_CAR.}

\item{est_type}{Type of IPW estimator, Hajek's stabilized IPW estimator or
the classical Horvitz-Thompson IPW estimator. Hajek's stabilized estimator
has some advantages in cases where the propensity score may suffer from
near-violations of the assumption of positivity. The default is to use the
classical Horvitz-Thompson IPW estimator.}
}
\description{
Compute the Coarsened-at-Random Projection of IPW Estimators
}
\keyword{internal}
