% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npipw.R
\name{ipw_eif}
\alias{ipw_eif}
\title{Compute the Efficient Influence Function of IPW Estimators}
\usage{
ipw_eif(a, y, gn, gn_eif = NULL, est_type = c("ht", "hajek"))
}
\arguments{
\item{a}{Observed values of the treatment.}

\item{y}{Observed values of the outcome.}

\item{gn}{Estimates of the propensity score gn used in computing the IPW
estimator. These same estimates may be reused in computing the EIF or be
replaced by the alternative \code{gn_eif}.}

\item{gn_eif}{Estimates of the propensity score used in the re-weighting
term (i.e., A / gn) of the EIF. The default is \code{NULL}, in which case
\code{gn} is used in the re-weighting term as well as in the IPW estimator.}

\item{est_type}{Type of IPW estimator, Hajek's stabilized IPW estimator or
the classical Horvitz-Thompson IPW estimator. Hajek's stabilized estimator
has some advantages in cases where the propensity score may suffer from
near-violations of the assumption of positivity. The default is to use the
classical Horvitz-Thompson IPW estimator.}
}
\description{
Compute the Efficient Influence Function of IPW Estimators
}
\keyword{internal}
