% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npipw.R
\name{ipw_est}
\alias{ipw_est}
\title{Compute IPW Estimators}
\usage{
ipw_est(a, y, gn, est_type = c("ht", "hajek"))
}
\arguments{
\item{a}{Observed values of the treatment.}

\item{y}{Observed values of the outcome.}

\item{est_type}{Type of IPW estimator, Hajek's stabilized IPW estimator or
the classical Horvitz-Thompson IPW estimator. Hajek's stabilized estimator
has some advantages in cases where the propensity score may suffer from
near-violations of the assumption of positivity. The default is to use the
classical Horvitz-Thompson IPW estimator.}

\item{gn_uhal}{Undersmoothed estimates of the propensity score gn.}
}
\description{
Compute IPW Estimators
}
\keyword{internal}
