% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{mse}
\alias{mse}
\title{Mean-squared error}
\usage{
mse(obs, estim)
}
\arguments{
\item{obs}{A \code{numeric} vector of observations.}

\item{estim}{A \code{numeric} vector of predictions.}
}
\description{
Mean-squared error
}
\keyword{internal}
