% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{nll}
\alias{nll}
\title{Risk under Cross-Entropy Loss}
\usage{
nll(obs, probs)
}
\arguments{
\item{obs}{A \code{numeric} vector of observed binary labels.}

\item{probs}{A \code{numeric} vector of predicted probabilities of class
assignment, of the form P(obs = 1).}
}
\description{
Risk under Cross-Entropy Loss
}
\keyword{internal}
